# Conflicts:
#   inc/ErrorHandler.hpp
#   inc/Time/TimeStamp.tpp

+++ test/Services/FileManagementServiceTests.cpp
@@ -289,7 +289,7 @@ TEST_CASE("File attributes report TM[23,4]", "[service][st23]") {
 	CHECK(report.readByte() == 'o');
 	CHECK(report.readByte() == 'r');
 	CHECK(report.readByte() == 't');
-	CHECK(report.readSint32() == fileContents.size());
+	CHECK(report.readSint32() == static_cast<int32_t>(fileContents.size()));
 	CHECK(report.readBoolean() == false);
 
 	fs::remove_all("st23");
